build: Migrate to uv project
