+++ pyproject.toml
@@ -1,0 +1,34 @@
+[project]
+name = "alecz-discord-bot"
+version = "0.1.0"
+description = "An Annoying Bot"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "aiodns>=3.6.1",
+    "aiohappyeyeballs==2.6.1",
+    "aiohttp==3.12.14",
+    "aiosignal==1.4.0",
+    "astunparse==1.6.3",
+    "attrs==23.2.0",
+    "braceexpand==0.1.7",
+    "brotli==1.2.0",
+    "cffi==1.16.0",
+    "click==8.1.7",
+    "colorama==0.4.6 ; sys_platform == 'win32'",
+    "discord-py==2.3.2",
+    "frozenlist==1.4.1",
+    "idna==3.7",
+    "import-expression==1.1.4",
+    "jishaku==2.5.2",
+    "multidict==6.0.5",
+    "orjson==3.10.2",
+    "propcache==0.3.2",
+    "pycares==4.9.0",
+    "pycparser==2.22",
+    "python-dotenv==1.0.1",
+    "six==1.16.0",
+    "typing-extensions==4.14.1 ; python_full_version < '3.13'",
+    "wheel==0.45.1",
+    "yarl==1.20.1",
+]
