--- requirements.in
@@ -1,4 +1,0 @@
-discord.py[speed]
-jishaku
-
-python-dotenv
