--- requirements.txt
@@ -1,62 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.12
-# by the following command:
-#
-#    pip-compile
-#
-aiodns==3.2.0
-    # via discord-py
-aiohappyeyeballs==2.6.1
-    # via aiohttp
-aiohttp==3.12.14
-    # via discord-py
-aiosignal==1.4.0
-    # via aiohttp
-astunparse==1.6.3
-    # via import-expression
-attrs==23.2.0
-    # via aiohttp
-braceexpand==0.1.7
-    # via jishaku
-brotli==1.2.0
-    # via discord-py
-cffi==1.16.0
-    # via pycares
-click==8.1.7
-    # via jishaku
-discord-py[speed]==2.3.2
-    # via -r requirements.in
-frozenlist==1.4.1
-    # via
-    #   aiohttp
-    #   aiosignal
-idna==3.7
-    # via yarl
-import-expression==1.1.4
-    # via jishaku
-jishaku==2.5.2
-    # via -r requirements.in
-multidict==6.0.5
-    # via
-    #   aiohttp
-    #   yarl
-orjson==3.10.2
-    # via discord-py
-propcache==0.3.2
-    # via
-    #   aiohttp
-    #   yarl
-pycares==4.9.0
-    # via aiodns
-pycparser==2.22
-    # via cffi
-python-dotenv==1.0.1
-    # via -r requirements.in
-six==1.16.0
-    # via astunparse
-typing-extensions==4.14.1
-    # via aiosignal
-wheel==0.45.1
-    # via astunparse
-yarl==1.20.1
-    # via aiohttp
